backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pymongo==4.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
Pillow==11.0.0
ultralytics==8.3.50
